move prepass to own file
